pyproject.toml
[tool.poetry]
name = "orama-challenge"
version = "0.1.0"
description = ""
authors = ["bgildson <gildson13@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
gunicorn = "^20.0.4"
marshmallow = "^3.7.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
flake8 = "^3.8.3"
flask-shell-ipython = "^0.4.1"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
